% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{activities_example}
\alias{activities_example}
\title{Example activity measurements}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3305 rows and 9 columns.}
\usage{
activities_example
}
\description{
Example activities for the variants in umpra_example
}
\keyword{datasets}
