% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barcode_by_id}
\alias{barcode_by_id}
\title{Example barcode_allele_df}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 471 rows and 2 columns.}
\usage{
barcode_by_id
}
\description{
Example barcode_allele_df, used for the example in count_barcodes(). Adapted
from Ulirsch et al., Cell, 2016.
}
\keyword{datasets}
