% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traditional_analysis_functions.R
\name{compute_activities}
\alias{compute_activities}
\title{Compute MPRA activities}
\usage{
compute_activities(mpra_data, rep_cutoff = 0.15,
  plot_rep_cutoff = TRUE, verbose = TRUE)
}
\arguments{
\item{mpra_data}{a data frame of MPRA data}

\item{rep_cutoff}{a DNA representation cutoff}

\item{plot_rep_cutoff}{a logical indicating whether or not to plot the DNA
barcode abundance and the applied representation cutoff}

\item{verbose}{logical indicating whether to print messages}
}
\value{
a data frame of activities
}
\description{
This function computes traditional MPRA activities, i.e. the
  log-ratio of RNA to DNA, for each barcode-RNA sample
}
\note{
the output is returned in a "tall" format, with sample_id's gathered
  into one column
}
\examples{
compute_activities(umpra_example)
}
