% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_analysis.R
\name{compute_kl_est}
\alias{compute_kl_est}
\title{Compute KL divergence estimate}
\usage{
compute_kl_est(remnant_set)
}
\arguments{
\item{remnant_set}{a data frame of count remnants (i.e. post sample/DNA
abundance normalization) with columns specifying the prior}
}
\value{
the Kullback-Leibler divergence between the empirical estimate and
  the prior
}
\description{
Approximate the KL divergence between a set of normalized counts
  and a prior. An empirical density
}
