% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cond_prior_example}
\alias{cond_prior_example}
\title{Example conditional prior}
\format{An object of class \code{list} of length 2.}
\usage{
cond_prior_example
}
\description{
Example conditional prior for the example variants taken from from Ulirsch et
al., Cell, 2016. This prior is the full, accurate prior using all 7000+
variants from that assay, so conditionl priors derived from the subsetted
example data included in malacoda will be different.
}
\keyword{datasets}
