% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{decode_errors}
\alias{decode_errors}
\title{Decode fastq}
\usage{
decode_errors(trimmed_fastqs, bc_set, temp_dir, other_fb_args = "")
}
\arguments{
\item{trimmed_fastqs}{a character string vector giving the paths to all the trimmed fastq
files to be decoded}

\item{bc_set}{a character string giving the path of the freebarcodes barcode
set file to use for decoding}

\item{temp_dir}{a character string giving the path to a directory to use for
temporary intermediate files}

\item{other_fb_args}{a character string giving other arguments to pass to
freebarcodes (besides barcode set and output directory)}
}
\description{
If one of the barcodes sets from
  \href{https://github.com/finkelsteinlab/freebarcodes}{freebarcodes} is
  used, this function can be used to decode the erroneous barcode reads.
}
\note{
See the
  \href{https://github.com/finkelsteinlab/freebarcodes}{freebarcodes github
  page} for more details on additional arguments, the decoding process, the
  freebarcodes package, and their original publication (Hawkins et al, Proc
  Natl Acad Sci, 2018).
}
