% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dropout_example}
\alias{dropout_example}
\title{Example CRISPR dropout screen counts}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 40 rows and 5 columns.}
\usage{
dropout_example
}
\description{
Example dropout screen counts for 10 randomly selected genes from "Optimized
libraries for CRISPR-Cas9 genetic screens with multiple modalities", Doench
et al., Nature Communications, 2018
}
\keyword{datasets}
