% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_posterior}
\alias{example_posterior}
\title{Example posterior}
\format{An object of class \code{stanfit} of dimension 75 x 4 x 33.}
\usage{
example_posterior
}
\description{
Example posterior for chr6:135426558 1/2 variant in Ulirsch et al., Cell,
2016. This example used too-few MCMC samples for the sake of keeping the file
size small.
}
\keyword{datasets}
