% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_result}
\alias{example_result}
\title{Example output of fit_mpra_model}
\format{a data frame of mpra modelling results}
\usage{
example_result
}
\description{
Example needed for the vignette
}
\keyword{datasets}
