% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fastq_examples}
\alias{fastq_examples}
\title{Example FASTQ data}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 200 rows and 12 columns.}
\usage{
fastq_examples
}
\description{
Example FASTQ entries, used for the example in count_barcodes(). Taken from
Ulirsch et al., Cell, 2016.
}
\keyword{datasets}
