% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_fitting.R
\name{find_prior_weights}
\alias{find_prior_weights}
\title{Find prior weights}
\usage{
find_prior_weights(given_id, scaled_annotations, dist_mat, min_dist_kernel,
  kernel_fold_change = 1.3, min_num_neighbors = 30)
}
\arguments{
\item{given_id}{a variant_id to get weights for}

\item{scaled_annotations}{a tall annotation data frame where the annotations
have been set to the same scale}

\item{dist_mat}{a distance matrix of euclidean distances between variants in
scaled annotation space}

\item{min_dist_kernel}{an initialization of the distance kernel at some tiny
value}

\item{kernel_fold_change}{the multiplier by which to iteratively increase the
distance kernel}

\item{min_num_neighbors}{the minimum number of neighbors which must make a
meaningful contribution to the weights before stopping}
}
\description{
For a given variant and annotation set (scaled) and distance
  matrix, get the weights of all other variants
}
\details{
"Meaningful contribution" to the weights means that the sum of the
  weights of the first min_num_neighbors of other variants must account for
  <= 99% of the total weight. This prevents some small number of extremely
  close neighbors from dominating the prior estimation later on.
}
