% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_fitting.R
\name{fit_cond_prior}
\alias{fit_cond_prior}
\title{Fit a informative conditional prior}
\usage{
fit_cond_prior(mpra_data, annotations, n_cores = 1,
  plot_rep_cutoff = TRUE, rep_cutoff = 0.15, min_neighbors = 30,
  kernel_fold_increase = 1.2)
}
\arguments{
\item{mpra_data}{a data frame of mpra data}

\item{annotations}{a data frame of annotations for the same variants in
mpra_data}

\item{n_cores}{number of cores to parallelize across}

\item{plot_rep_cutoff}{logical indicating whether to plot the representation
cutoff used}

\item{rep_cutoff}{fraction indicating the depth-adjusted DNA count quantile
to use as the cutoff}

\item{min_neighbors}{The minimum number of neighbors in annotation spcae that
must contribute to prior estimation}

\item{kernel_fold_increase}{The amount to iteratively increase kernel width
by when estimating conditional priors. Smaller values (closer to 1) will
yield more refined priors but take longer.}
}
\value{
A data frame with a weighted conditional prior for each variant
}
\description{
Use informative annotations to bias prior estimation towards
  alleles that show similar annotations
}
\details{
The
}
