% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_fitting.R
\name{fit_marg_prior}
\alias{fit_marg_prior}
\title{Fit a marginal prior}
\usage{
fit_marg_prior(mpra_data, n_cores = 1, plot_rep_cutoff = TRUE,
  rep_cutoff = 0.15)
}
\arguments{
\item{mpra_data}{a data frame of mpra data}

\item{n_cores}{number of cores to parallelize across}

\item{plot_rep_cutoff}{logical indicating whether to plot the representation cutoff used}

\item{rep_cutoff}{fraction indicating the depth-adjusted DNA count quantile to use as the cutoff}
}
\description{
Fit a marginal prior
}
