% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_functions.R
\name{fit_nb}
\alias{fit_nb}
\title{Fit a negative binomial}
\usage{
fit_nb(input_dat, nb_init = c(10, 1))
}
\arguments{
\item{input_dat}{a data frame with one column called "counts"}

\item{nb_init}{optional vector of mean, dispersion initialization point}
}
\description{
Fit a negative binomial distribution to a set of counts
}
