% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_fitting.R
\name{generate_distance_matrix}
\alias{generate_distance_matrix}
\title{Generate a distance matrix from a matrix of annotations}
\usage{
generate_distance_matrix(annotations, log_distance = FALSE,
  scale_annotations = TRUE)
}
\arguments{
\item{annotations}{an n x d data frame of annotations}

\item{log_distance}{a logical indicating to use the log1p of the distances (TRUE) or the raw euclidean distances (FALSE)}

\item{scale_annotations}{logical indicating whether to base::scale to center and scale annotations}
}
\description{
Given an nxd matrix of variant annotations, produce an nxn distance matrix
describing the inter-variant distances in annotation space
}
