% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_analysis.R
\name{get_kl_divergences}
\alias{get_kl_divergences}
\title{Get Kullback-Leibler Divergences}
\usage{
get_kl_divergences(mpra_data, cond_prior, marg_prior, n_cores, verbose = TRUE)
}
\arguments{
\item{mpra_data}{a data frame of mpra data}

\item{cond_prior}{a conditional prior (see fit_cond_prior())}

\item{marg_prior}{a marginal prior (see fit_marg_prior())}

\item{n_cores}{number of cores for parallelization}

\item{verbose}{logical indicating whether to print messages}
}
\value{
a data frame giving the KL between the observed normalized counts and
  the two priors for each allele of each variant_id
}
\description{
Assess the prior improvement using Kullback-Leibler divergence
}
\note{
The KL divergences are approximations because gamma kernel density
  estimation is used to obtain the empirical distribution of the normalized
  counts.
}
