% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_label_y}
\alias{get_label_y}
\title{Get ratio label y}
\usage{
get_label_y(ratio_values)
}
\arguments{
\item{ratio_values}{a vector of ratio values to get a relative position from}
}
\description{
helper function for plot_prior_ratios
}
