% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_analysis.R
\name{get_prior_ratios}
\alias{get_prior_ratios}
\title{Evaluate prior ratios of MLE estimates}
\usage{
get_prior_ratios(mpra_data, marg_prior, cond_prior, n_cores = 1)
}
\arguments{
\item{mpra_data}{a data frame of MPRA data}

\item{marg_prior}{a marginal prior}

\item{cond_prior}{a conditional prior}

\item{n_cores}{number of cores to use in parallel}
}
\value{
a data frame of MLE mean and dispersion parameters by variant_id,
  sample_id, and barcode giving prior ratio for each
}
\description{
This function evaluates prior ratios of MLE estimates under
  user-supplied conditional and marginal prior distributions. If this ratio
  is above one (or equivalently if the log is above 0), this implies that the conditional prior improves upon the
  marginal prior for that particular parameter.
}
\details{
the inputs \code{marg_prior} and \code{cond_prior} can be taken
  directly as the outputs of fit_marg_prior and fit_cond_prior.

  This output is returned as the difference of log densities.
}
\note{
The priors for DNA counts are always the same (since we have no prior
  knowledge of DNA counts)

  The current maximum likelihood estimates are sub-par and will be improved. Interpret with caution.
}
\examples{
marg_prior = fit_marg_prior(umpra_example)
cond_prior = fit_cond_prior(mpra_data = umpra_example,
                            annotations = u_deepsea,
                            min_neighbors = 10)
get_prior_ratios(umpra_example, marg_prior, cond_prior)
}
