% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{get_sample_correlations}
\alias{get_sample_correlations}
\title{Get MPRA sample correlations}
\usage{
get_sample_correlations(mpra_data)
}
\arguments{
\item{mpra_data}{a data frame of mpra data}
}
\value{
a data frame of pairwise correlations
}
\description{
Compute the correlations of MPRA sequencing samples. This can be
  a helpful as a QC diagnostic metric. Libraries of the same type (DNA or RNA) should
  correlate highly with one another
}
\details{
the output is ready to be presented to plot_mpra_correlations
}
