% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_fitting.R
\name{get_well_represented}
\alias{get_well_represented}
\title{Get well represented barcodes}
\usage{
get_well_represented(mpra_data, sample_depths, rep_cutoff,
  plot_rep_cutoff = FALSE, verbose = TRUE)
}
\arguments{
\item{mpra_data}{a data frame of MPRA data}

\item{sample_depths}{a data frame of sample depths}

\item{rep_cutoff}{a representation cutoff}

\item{plot_rep_cutoff}{logical indicating whether to plot the DNA representation distribution with the input rep_cutoff indicated}
}
\description{
Identify barcodes well-represented in DNA samples from input MPRA data.
}
