% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_fitting.R
\name{increase_regularization}
\alias{increase_regularization}
\title{Increase a malacoda prior object's regularization}
\usage{
increase_regularization(prior_object, increase_factor = 2)
}
\arguments{
\item{prior_object}{a marginal prior object returned by fit_marg_prior()}

\item{increase_factor}{a number indicating the factor by which to increase}
}
\value{
A malacoda prior object with increase
}
\description{
This function takes a malacoda prior object and increases the
  regularization on transcription shift by a set factor.
}
\details{
This function multiplies the alpha and beta parameters of the gamma
  priors on the mean for the RNA counts in order to keep the mean the same
  while ascribing less prior density to extreme values, thus effectively
  increasing the shrinkage applied to transcription shift while evaluating
  the posterior.
}
\examples{
increase_regularization(marg_prior_example)
}
\seealso{
\code{\link{fit_marg_prior}} \code{\link{sample_from_prior}}
  \code{\link{summarise_prior_samples}}
}
