% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{malacoda_intervals}
\alias{malacoda_intervals}
\title{Plot a variant's posterior intervals}
\usage{
malacoda_intervals(
  sampler_res,
  inner_probs = c(0.1, 0.9),
  outer_probs = c(0.025, 0.975)
)
}
\arguments{
\item{sampler_res}{A stanfit object for one variant}

\item{inner_probs}{quantiles of the posterior for the inner bar}

\item{outer_probs}{quantiles of the posterior for the outer bar}
}
\description{
Given a result object from the malacoda model, this function
  plots the posterior intervals for all the modeled parameters. The outer
  black bars represent the limits of the 95% interval, while the inner blue
  bar represents the 80% interval (by default). The black dots are the
  posterior means.
}
\details{
The sampler_res object for a given variant is saved into the out_dir
  argument of fit_mpra_model. The barcode concentrations (i.e. DNA means) are
  numbered in the same order in which they are given to fit_mpra_model, with
  poorly represented barcodes removed entirely.
}
\examples{
malacoda_intervals(example_posterior)
}
\seealso{
fit_mpra_model get_well_represented
}
