% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marg_prior_example}
\alias{marg_prior_example}
\title{Example marginal prior}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 6 columns.}
\usage{
marg_prior_example
}
\description{
Example marginal prior derived from MPRA data from Ulirsch et al., Cell,
2016.
}
\keyword{datasets}
