% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{match_barcodes}
\alias{match_barcodes}
\title{Match barcodes for a fastq}
\usage{
match_barcodes(barcode_allele_df, trimmed_fastq, temp_dir)
}
\description{
Call fastx_barcode_splitter on one trimmed fastq
}
