% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{mpra_tile_plot}
\alias{mpra_tile_plot}
\title{Make a MPRA tile plot}
\usage{
mpra_tile_plot(
  variant_id,
  variant_counts,
  show_barcodes = FALSE,
  composite = FALSE,
  sampler_res = NULL,
  prob = 0.5,
  prob_outer = 0.95
)
}
\arguments{
\item{variant_id}{variant ID}

\item{variant_counts}{a data frame of counts for one variant}

\item{show_barcodes}{a logical indicating whether to show the barcodes on the
left side of the tile plot}

\item{composite}{a logicial indicating whether or not to include the extra
plot components on the right sides}

\item{sampler_res}{A stanfit object for one variant}

\item{prob}{if composite = TRUE, a fraction between 0 and 1 indicating the
posterior probability mass to include in the inner intervals.}

\item{prob_outer}{if composite = TRUE, a fraction between 0 and 1 indicating
the posterior probability mass to include in the outer intervals.}
}
\description{
Make a tile plot of the MPRA counts for a given variant
}
\details{
The barcodes in the output tile plot are sorted according the their
  average DNA representation.

  The barcodes in variant_counts need to be in the same order that they were
  presented to the malacoda sampler. malacoda tries to preserve the ordering
  present in the original mpra_data dataset (not including removed poorly
  represented barcodes), but if you're unsure you can pull from the
  variant_data column in the analysis_res object. Currently it's necessary
  that the allele be encoded by 'alt' and 'ref'.

  Setting composite = TRUE will include three additional components: a column
  of tiles whose color indicates the log(mean(RNA)/ mean(DNA)) by row, and
  posterior interval plots for the variant's activities and transcription
  shift.
}
