% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_dna_representation}
\alias{plot_dna_representation}
\title{Plot DNA representation}
\usage{
plot_dna_representation(dna_df, rep_cutoff)
}
\arguments{
\item{dna_df}{a dataframe of depth-normalized DNA abundances}

\item{rep_cutoff}{a representation cutoff}
}
\value{
a DNA representation ggplot
}
\description{
Plot DNA representation histogram
}
\note{
see the first command in the source code of get_well_represented() for
  how to generate dna_df
}
