% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_mpra_correlations}
\alias{plot_mpra_correlations}
\title{Plot MPRA sample correlations}
\usage{
plot_mpra_correlations(sample_correlations)
}
\arguments{
\item{sample_correlations}{a data frame of pairwise sample correlations}
}
\value{
a ggplot showing the pairwise sample correlations
}
\description{
Plot MPRA sample correlations
}
\details{
This function visualizes the pairwise correlations of count samples
  in MPRA data. This can be a useful QC metric - samples should correlate
  highly with other samples of the same type (DNA or RNA). If ALL samples
  correlate highly with one another, this can indicate DNA contamination in
  the RNA libraries.
}
\note{
Get \code{sample_correlations} from get_mpra_correlations
}
