% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_prior_samples}
\alias{plot_prior_samples}
\title{Plot prior samples}
\usage{
plot_prior_samples(prior_draws)
}
\arguments{
\item{prior_draws}{a data frame of prior samples}
}
\value{
a ggplot visualizing the prior density on TS
}
\description{
Visualize a malacoda prior on transcription shift
}
\examples{
prior_draws = sample_from_prior(marg_prior_example, n_samp = 2000)
plot_prior_samples(prior_draws)
}
\seealso{
\code{\link{sample_from_prior}}
}
