% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_ratio_hexs}
\alias{plot_ratio_hexs}
\title{Create ratio by transcription shift hexbins}
\usage{
plot_ratio_hexs(prior_ratios, activities, y_limits = c(-1, 1))
}
\arguments{
\item{prior_ratios}{a dataframe of barcode activities}

\item{activities}{a dataframe of barcode activities}

\item{y_limits}{}
}
\description{
Create ratio by transcription shift hexbins
}
\details{
prior_ratios can be produced by get_prior_ratios() and activities
  can be produced by compute_activities
}
