% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{posterior_beeswarm}
\alias{posterior_beeswarm}
\title{Create a posterior beeswarm plot}
\usage{
posterior_beeswarm(sampler_result, activities, color_by_sample = FALSE)
}
\arguments{
\item{sampler_result}{a stanfit object}

\item{activities}{a dataframe of activities (only for the same variant as sampler_result)}

\item{color_by_sample}{optional logical indicating whether to color dots by
sample}
}
\value{
a posterior beeswarm ggplot object
}
\description{
Create a combined beeswarm & activity posterior plot
}
\note{
sampler_result objects are written by fit_mpra_model to the out_dir argument for each variant_id
}
