% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ratios_example}
\alias{ratios_example}
\title{Example prior ratios object}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7086 rows and 13 columns.}
\usage{
ratios_example
}
\description{
Example prior ratios for the conditional and marginal priors trained on the
example dataset. See the example in ?plot_prior_ratios to see how it is computed.
}
\keyword{datasets}
