% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traditional_analysis_functions.R
\name{run_activity_tests}
\alias{run_activity_tests}
\title{Run MPRA activity tests}
\usage{
run_activity_tests(mpra_activities, test_type = "t",
  compute_FDR = TRUE, plot_p_hist = TRUE)
}
\arguments{
\item{mpra_activities}{a "tall" data frame of MPRA activities}

\item{test_type}{a string indicating the type of test to use to compare
activities, either 't' or 'U'}

\item{compute_FDR}{logical indicating whether or not to add on a column of
FDR q-values}

\item{plot_p_hist}{logical indicating whether or not to plot a histogram of p-values}
}
\value{
a data frame of test statistics
}
\description{
Given a set of MPRA activity measurements, run significance
  tests to see if there's a difference in activity tests between alleles
}
\details{
Compute activity tests by MPRA variant, comparing the activity
  measurements between alleles across all sequencing samples.
}
\note{
These tests assume that after depth correction, MPRA activities are
  identically distributed across sequencing samples. Thus activities for the
  same barcodes across different samples are considered as independent
  measurements.
}
\examples{
run_activity_tests(activities_example)
}
