% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_analysis.R
\name{score_annotation}
\alias{score_annotation}
\title{Score an annotation source}
\usage{
score_annotation(
  mpra_data,
  annotations,
  nb_params = NULL,
  conditional_prior = NULL,
  marginal_prior = NULL
)
}
\arguments{
\item{mpra_data}{a data frame of MPRA data}

\item{annotations}{a data frame containing annotations of the same alleles in mpra_data}

\item{nb_params}{an optional data frame of pre-fit maximum likelihood estimates for allele parameters}

\item{conditional_prior}{an optional data frame containing pre-fit conditional priors}

\item{marginal_prior}{an optional data frame containing a pre-fit marginal prior}
}
\value{
A fraction between 0 and 1 indicating the number of parameter
  estimates that have higher density under the conditional prior than under
  the marginal prior.
}
\description{
This function takes MPRA data and an annotation source (or
  alternatively pre-fit priors) and checks how well the conditional prior
  improves upon the marginal prior by a prior ratio at the maximum likelihood
  estimates. A conditional:marginal ratio > 1 indicates that the conditional
  prior does better. A higher fraction of alleles for which this is true
  indicates an improved conditional prior.
}
\note{
Note that even with near-perfect annotations or conditional priors, the
  fraction of parameter estimate ratios above 1 will not approach 1. It will
  instead approach some difficult-to-determine limit defined by the noise in
  the assay. Thus this functionality should only be used to COMPARE
  annotation sources, rather than make discrete claims / measurements about
  individual annotation sources.
}
