% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_functions.R
\name{summarise_cond_prior}
\alias{summarise_cond_prior}
\title{Summarise a conditional prior}
\usage{
summarise_cond_prior(cond_prior, n_cores = 1, n_samp = 50000)
}
\arguments{
\item{cond_prior}{a conditional prior object}

\item{n_cores}{number of cores for parallelization}

\item{n_samp}{number of prior simulation draws}
}
\value{
a data frame of RNA priors with prior simulation summary statistics
}
\description{
This function runs Monte Carlo simulations for each variant's
  conditional mean priors.
}
\details{
This is used to ensure that the prior is not TOO specific. While we
  want the prior distribution to accurately reflect prior beliefs on a given
  variant's effects, it would be undesirable for the prior to pre-specify
  that the variant is functional. If the \code{prior_is_func} column in the
  output of this function is all FALSE, this means that no variant is \emph{a
  priori} functional.

  If one does get a conditional prior that has variants that are \emph{a
  priori} functional, this can be addressed by increasing the
  \code{min_neighbors} argument in \code{fit_cond_prior}. See the
  documentation of that function for details.

  Currently this function only simulates the mean parameters as in practice
  the dispersion parameters don't vary systematically between alleles and/or
  variants.
}
\note{
simulations for individual variants can be obtained with
  \code{summarise_one_prior()} and \code{sample_from_prior()}
}
\examples{
summarise_cond_prior(cond_prior_example, n_samp = 1000, n_cores = 1)
}
