% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_functions.R
\name{summarise_one_prior}
\alias{summarise_one_prior}
\title{Summarise one RNA prior}
\usage{
summarise_one_prior(prior_df, n_samp)
}
\arguments{
\item{prior_df}{a single RNA prior}

\item{n_samp}{number of prior draws used in simulation}
}
\value{
a data frame with summary statistics for the input prior
}
\description{
Sample from one RNA prior and compute summary statistics
}
