% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_functions.R
\name{summarise_prior_samples}
\alias{summarise_prior_samples}
\title{Summarise prior samples}
\usage{
summarise_prior_samples(sim_df)
}
\arguments{
\item{sim_df}{a data frame of simulation draws}
}
\value{
a row data frame with columns of summary statistics
}
\description{
Compute summary statistics for input prior simulation draws
}
