% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{trim_and_filter}
\alias{trim_and_filter}
\title{Trim and Filter a FASTQ}
\usage{
trim_and_filter(fastq, temp_dir, quality_cutoff, bc_start, bc_end)
}
\arguments{
\item{fastq}{path to a fastq file}

\item{temp_dir}{a character string giving the path to a directory to use for
temporary intermediate files}

\item{quality_cutoff}{an integer indicating the quality cutoff to be used}

\item{bc_start}{an integer indicating the position in the read where the
barcode starts}

\item{bc_end}{an integer indicating the position in the read where the
barcode ends}
}
\description{
Use the FASTX-Toolkit to trim out the barcodes from the FASTQs by position, then quality filter them.
}
