% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{u_deepsea}
\alias{u_deepsea}
\title{Example MPRA annotation set}
\format{a data frame containing variant_id and annotation columns}
\source{
\url{http://deepsea.princeton.edu/}
}
\usage{
u_deepsea
}
\description{
DeepSea v0.94 K562 DNase hypersensitivity estimates for the variants in Ulirsch et
al., Cell, 2016.
}
\keyword{datasets}
