% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_functions.R
\name{fit_dropout_model}
\alias{fit_dropout_model}
\title{Fit a Bayesian model of dropout CRISPR screen data}
\usage{
fit_dropout_model(dropout_data, out_dir = NULL, n_cores = 1,
  tot_samp = 10000, n_warmup = 500, n_chains = 4, rep_cutoff = 0.1,
  plot_rep_cutoff = TRUE)
}
\arguments{
\item{dropout_data}{a data frame of dropout data. See details for column
requirements.}

\item{out_dir}{path to output directory}

\item{n_cores}{number of cores to utilize}

\item{tot_samp}{total number of MCMC draws to take, spread evenly across
chains}

\item{n_warmup}{total number of warmup draws to take from each MCMC chain}

\item{n_chains}{number of MCMC chains to run in each sampler}

\item{rep_cutoff}{a representation cutoff quantile (0 to 1)}

\item{plot_rep_cutoff}{logical indicating whether to plot the representation
cutoff histogram}
}
\value{
a data frame of input counts, fit and model statistics for the
  log-fold-change for each input gene.
}
\description{
This function fits a Bayesian model of survival/dropout CRISPR
  screen data. It uses a negative binomial to model the input and output
  counts of gRNAs, adjusting the results appropriately to account for
  sequencing depth.
}
\details{
\code{dropout_data} requires the following columns:
  \itemize{\item{gene_id - character column giving a unique identifier for
  each gene} \item{gRNA - character column giving identifiers for individual
  gRNAs (usually the gRNA sequence itself)} \item{input count columns -
  columns of sequencing counts of the input gRNA library. Multiple columns
  for sequencing replicates are allowed (which require unique identifiers).
  Column names must contain the string "input".} \item{output count columns -
  columns of sequencing counts of gRNAs in the output libraries. Multiple
  columns allowed (which in turn require unique names). Column name must
  contain the string "output".}}
}
\note{
Currently this function only supports marginal priors. If you want to
  use grouped/conditional priors, contact the malacoda developers.

  The \code{gene_data} column in the output contains only the gRNAs that
  passed the representation cutoff.
}
\examples{
# This example uses too-few MCMC samples for the sake of run time. Convergence will be poor.

fit_dropout_model(dropout_data = dropout_example,
n_cores = 1,
tot_samp = 50,
n_warmup = 5)
}
