% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_fitting.R
\name{fit_marg_prior}
\alias{fit_marg_prior}
\title{Fit a marginal prior}
\usage{
fit_marg_prior(mpra_data, n_cores = 1, plot_rep_cutoff = TRUE,
  rep_cutoff = 0.15, sample_depths, well_represented, verbose = TRUE)
}
\arguments{
\item{mpra_data}{a data frame of mpra data}

\item{n_cores}{number of cores to parallelize across}

\item{plot_rep_cutoff}{logical indicating whether to plot the representation
cutoff used}

\item{rep_cutoff}{fraction indicating the depth-adjusted DNA count quantile
to use as the cutoff}

\item{sample_depths}{optional inputs to allow passing in sample_depths and
well_represented objects}

\item{well_represented}{optional inputs to allow passing in sample_depths and
well_represented objects}

\item{verbose}{logical indicating whether to print messages}
}
\description{
Fit a marginal prior
}
\examples{
marg_prior = fit_marg_prior(umpra_example,
 rep_cutoff = .15,
 plot_rep_cutoff = TRUE,
 n_cores = 1)
}
