% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_functions.R
\name{fit_mpra_model}
\alias{fit_mpra_model}
\title{Fit a Bayesian MPRA model}
\usage{
fit_mpra_model(mpra_data, annotations = NULL, group_df = NULL, out_dir,
  save_nonfunctional = FALSE, priors = NULL, n_cores = 1,
  n_chains = 4, tot_samp = 10000, n_warmup = 500, vb_pass = TRUE,
  vb_prob = 0.8, ts_hdi_prob = 0.95, ts_rope = NULL,
  rep_cutoff = 0.15)
}
\arguments{
\item{mpra_data}{a data frame of MPRA data with 1 column called variant_id,
an allele column, a barcode column, and additional columns per sequencing
sample. Each row is for a single barcode.}

\item{annotations}{an optional data frame of annotations with identical
variant_ids and an arbitrary number of functional annotations. If omitted,
the prior for a given variant is influenced by all other variants in the
assay equally.}

\item{group_df}{an optional data frame giving group identity by variant_id in mpra_data}

\item{out_dir}{path to output directory}

\item{save_nonfunctional}{logical indicating whether or not to save the
sampler results for variants identified as non-functional}

\item{priors}{optional objects provided by either fit_marg_prior() or
fit_cond_prior.}

\item{n_cores}{number of cores across which to parallelize variant MPRA
samplers}

\item{n_chains}{number of MCMC chains to run in each sampler}

\item{tot_samp}{total number of MCMC draws to take, spread evenly across
chains}

\item{n_warmup}{total number of warmup draws to take from each MCMC chain}

\item{vb_pass}{logical indicating whether to use a variational first pass}

\item{vb_prob}{numeric 0 - 1 indicating probability mass to use as a TS HDI
for identifying "promising" candidates for MCMC followup}

\item{ts_hdi_prob}{probability mass to include in the highest density
interval on transcription shift to call MPRA-functional variants.}

\item{ts_rope}{optional length 2 numeric vector describing the boundaries of
the transcription shift region of practical equivalence}

\item{rep_cutoff}{a representation cutoff quantile (0 to 1)}
}
\value{
a data frame with a row for each variant_id that specificies the
  posterior mean TS, upper and lower HDI bounds, a binary call of functional
  or non-functional, and other appropriate outputs
}
\description{
This function fits a negative-binomial based Bayesian model to
  MPRA data. Optional annotations can be included to allow for more
  informative conditional priors.
}
\details{
\code{mpra_data} must contain the following groups of columns:
  \itemize{ \item{variant_id} \item{allele - either 'ref' or 'alt'}
  \item{barcode - a unique index sequence for that row (ideally the same
  barcode used in the assay)} \item{at least one column of MPRA counts whose
  column name(s) matches 'DNA'} \item{at least one column of MPRA counts
  whose column name(s) matches 'RNA'} }

  \code{annotations} must contain the same variant_id's used in mpra_data.
  Additional columns are used as informative predictors: when estimating the
  priors for one variant, other variants with similar annotations will be
  upweighted in the prior-fitting process.

  If \code{priors} is provided, any annotations input will be ignored. This
  can be useful when you want to fit models again without having to spend
  time re-fitting the priors.

  Sampler results will be saved to out_dir. By default, only the sampler
  results for variants identified as MPRA-functional will be saved. This
  behavior can be changed by setting \code{save_nonfunctional} to TRUE.

  We've set the sampler parameters (n_chains to n_warmup) to values that work
  reasonably well at reasonable speeds for typical MPRA data on typical
  hardware. Final analyses and/or models fit to larger MPRA experiments will
  likely want to increase n_chains and tot_samp considerably to ensure
  precise convergence.

  \code{vb_pass} indicates whether to use a first pass variational check to
  see if a given variant is worth running the MCMC sampler. It does this by
  checking if a 40% HDI on the variational transcription shift posterior
  excludes 0. This speeds up posterior evaluation considerably, but gives
  approximate results. If \code{vb_pass} is set to FALSE, all variants get
  MCMC.

  \code{ts_rope} can be used to define a "Region Of Practical Equivalence"
  for transcription shift. This is some small-ish region around 0 where
  observed posterior samples are "practically equivalent" to 0. Enabling this
  option will return the fraction of transcription shift posterior samples
  that fall within the defined ROPE along with the usual model outputs. If
  this fraction is small, one can say that there is very little posterior
  belief that the variant's transcription shift is practically equivalent to
  0. If used, the user must be cognizant of defining the region in accordance
  with observed noise and effect size levels. Note that the output ROPE
  fractions ARE NOT p-values.

  Barcodes below the \code{rep_cutoff} quantile of representation in the DNA
  pools are discarded.
}
\note{
Sampler results for individual variants will be saved to the specified
  out_dir as they can be several megabytes each
}
