% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_ratio_hexs}
\alias{plot_ratio_hexs}
\title{Create ratio by transcription shift hexbins}
\usage{
plot_ratio_hexs(prior_ratios, model_result, y_limits = c(-0.5, 0.5))
}
\arguments{
\item{prior_ratios}{a dataframe of barcode prior_ratios}

\item{model_result}{a dataframe of model results}

\item{y_limits}{y limits of the plot}
}
\description{
Create ratio by transcription shift hexbins
}
\details{
prior_ratios can be produced by get_prior_ratios() and model_result
  can be obtained from fit_mpra_model()
}
\note{
The default y_limits cut off a small fraction of points where one prior
  or the other does vastly better. It can be set to larger values to avoid
  this behavior.
marg_prior = fit_marg_prior(umpra_example)
cond_prior = fit_cond_prior(mpra_data = umpra_example,
                            annotations = u_deepsea,
                            min_neighbors = 10)
ratio_df = get_prior_ratios(umpra_example, marg_prior, cond_prior)
example_variants = c("11_8839229_1-2", "15_75303554_2-3", "1_203652141_2-3")

examples_to_evaluate = umpra_example[umpra_example$variant_id %in% example_variants,]
example_result = fit_mpra_model(mpra_data = examples_to_evaluate,
 priors = marg_prior,
vb_pass = FALSE,
tot_samp = 100,
n_warmup = 10) # Likewise, n_warmup should be >500

plot_ratio_hexs(ratio_df, example_result)
}
