% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_functions.R
\name{sample_from_prior}
\alias{sample_from_prior}
\title{Sample from prior}
\usage{
sample_from_prior(prior_df, n_samp = 5000)
}
\arguments{
\item{prior_df}{a data frame of the prior for one variant}

\item{n_samp}{number of prior simulation draws to pull}
}
\value{
a data data frame of activity draws and the simulated transcription
  shift at each draw
}
\description{
Randomly draw activities levels and transcription shifts from
  the prior
}
\details{
This function draws \code{n_samp} draws from the prior for one
  variant.
}
\examples{
sample_from_prior(marg_prior_example, n_samp = 2000)
}
