% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{umpra_example}
\alias{umpra_example}
\title{Example MPRA counts from Ulirsch et al., 2016}
\format{a data frame giving variant_id, allele identifier, barcode, and
  sequencing counts for the two DNA libraries and 6 control RNA libraries}
\source{
\url{https://www.bloodgenes.org/RBC_MPRA}
}
\usage{
umpra_example
}
\description{
MPRA counts for 25 randomly chosen variants from Ulirsch et al., Cell, 2016.
Original URL seems to be down as of 10/2/2018.
}
\keyword{datasets}
